Fix http connection reuse
At Datameer, we have had failing communication between client and server
when the client reused HTTP connections. When I looked at the code, I
noticed that the author(s) of the ConnectionKeepAliveStrategy seemed to
think that it does what is actually the purpose of a
ConnectionReuseStrategy. The comments within getKeepAliveDuration(..)
suggest that they thought that by returning -1 they disable connection
reuse (it actually means "keep the connection alive indefinitely").
However, I'm not sure if that is the real root cause of our issues
regarding failing communication between client and server when connections
are reused, so I designed the new code so that connections are never reused
unless told to via property.